combine large solar project rules
