+++ workflow/scripts/prepare_large_solar.py
@@ -1,4 +1,8 @@
-"""Processing of the Tranzition Zero Solar Asset Mapper (TZ-SAM) dataset."""
+"""Processing of large-scale solar powerplant datasets.
+
+- Tranzition Zero Solar Asset Mapper (TZ-SAM) dataset.
+- GEM Global Solar Power Tracker (GEM-GSPT) dataset.
+"""
 
 import sys
 from typing import TYPE_CHECKING, Any
@@ -52,7 +56,7 @@ def fill_tz_with_gem(
     return filled_tz_sam
 
 
-def get_gem_mismatch(
+def get_gem_v_tz_mismatch(
     gem_df: gpd.GeoDataFrame,
     tz_sam_df: gpd.GeoDataFrame,
     valid_status: list[str] | None = None,
@@ -139,7 +143,7 @@ def prepare_solar_utility_pv(
     )
 
     filled_tz_df = fill_tz_with_gem(gem_df, tz_df)
-    gem_mismatch_df = get_gem_mismatch(
+    gem_mismatch_df = get_gem_v_tz_mismatch(
         gem_df,
         tz_df,
         valid_status=["announced", "pre-construction", "construction", "retired"],
@@ -152,6 +156,37 @@ def prepare_solar_utility_pv(
     return schema.validate(utility_pv)
 
 
+def prepare_solar_csp(
+    gem_gspt_path: str, tech_name: str, dc_ac_ratio: float = 1
+) -> gpd.GeoDataFrame:
+    """Obtain CSP power locations using GEM-GSPT data."""
+    raw_df = gem.read_gem_dataset(gem_gspt_path, gem.GEM_GSPT_SHEETS)
+    raw_df = raw_df[raw_df["technology_type"] == "Solar Thermal"]
+
+    csp_df = gpd.GeoDataFrame(
+        {
+            "powerplant_id": _utils.get_combined_text_col(
+                raw_df, ["gem_location_id", "gem_phase_id"], prefix="GEM_"
+            ),
+            "name": _utils.get_combined_text_col(
+                raw_df, ["project_name", "phase_name"]
+            ),
+            "category": "solar",
+            "technology": tech_name,
+            "output_capacity_mw": gem.output_capacity_mw_gspt(
+                raw_df, dc_ac_ratio, "AC"
+            ),
+            "start_year": gem.year_col(raw_df, "start"),
+            "end_year": gem.year_col(raw_df, "end"),
+            "status": gem.status_col(raw_df),
+            "geometry": _utils.get_point_col(raw_df, "longitude", "latitude"),
+        }
+    ).reset_index(drop=True)
+    schema = _schemas.build_schema({"csp": tech_name}, "prepare")
+    return schema.validate(csp_df)
+
+
+
 def main() -> None:
     """Main snakemake process."""
     sys.stderr = open(snakemake.log[0], "w")
@@ -161,7 +196,10 @@ def main() -> None:
         snakemake.params.utility_pv_name,
         snakemake.params.dc_ac_ratio,
     )
-    utility_pv_gdf.to_parquet(snakemake.output.path)
+    csp_gdf = prepare_solar_csp(snakemake.input.gem_gspt, snakemake.params.csp_name)
+
+    utility_pv_gdf.to_parquet(snakemake.output.utility_pv)
+    csp_gdf.to_parquet(snakemake.output.csp)
 
 
 if __name__ == "__main__":
